requirements.txt
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.35
langchain-huggingface==0.3.1
langchain-chroma==0.2.6
sentence-transformers==5.1.1
tf_keras==2.20.1
python-dotenv==1.1.1
